restructured grantee components in permissions
