+++ graylog2-web-interface/src/components/permissions/Grantee/GranteeIcon.tsx
